[IMP] account: account_type_selection widget owl migration
