+++ addons/account/static/src/js/account_selection.js
@@ -1,31 +1,23 @@
 /** @odoo-module **/
 
-import { qweb, _t } from 'web.core';
-import fieldRegistry from 'web.field_registry';
-import { FieldSelection } from 'web.relational_fields';
+import { registry } from "@web/core/registry";
+import { SelectionField } from "@web/views/fields/selection/selection_field";
 
-
-export const HierarchySelection = FieldSelection.extend({
-
-    init: function () {
-        this._super.apply(this, arguments);
-        this.hierarchyGroups = [
-            {name: _t('Balance Sheet')},
-            {name: _t('Assets'), children: this.values.filter(x => x[0] && x[0].startsWith('asset'))},
-            {name: _t('Liabilities'), children: this.values.filter(x => x[0] && x[0].startsWith('liability'))},
-            {name: _t('Equity'), children: this.values.filter(x => x[0] && x[0].startsWith('equity'))},
-            {name: _t('Profit & Loss')},
-            {name: _t('Income'), children: this.values.filter(x => x[0] && x[0].startsWith('income'))},
-            {name: _t('Expense'), children: this.values.filter(x => x[0] && x[0].startsWith('expense'))},
-            {name: _t('Other'), children: this.values.filter(x => x[0] && x[0] == 'off_balance')},
+export class AccountTypeSelection extends SelectionField {
+    get hierarchyOptions() {
+        const opts = this.options;
+        return [
+            { name: this.env._t('Balance Sheet') },
+            { name: this.env._t('Assets'), children: opts.filter(x => x[0] && x[0].startsWith('asset')) },
+            { name: this.env._t('Liabilities'), children: opts.filter(x => x[0] && x[0].startsWith('liability')) },
+            { name: this.env._t('Equity'), children: opts.filter(x => x[0] && x[0].startsWith('equity')) },
+            { name: this.env._t('Profit & Loss') },
+            { name: this.env._t('Income'), children: opts.filter(x => x[0] && x[0].startsWith('income')) },
+            { name: this.env._t('Expense'), children: opts.filter(x => x[0] && x[0].startsWith('expense')) },
+            { name: this.env._t('Other'), children: opts.filter(x => x[0] && x[0] === 'off_balance') },
         ];
-    },
-
-    _renderEdit: function () {
-        this.$el.empty();
-        this.$el.append(qweb.render('accountTypeSelection', {widget: this}));
-        this.$el.val(JSON.stringify(this._getRawValue()));
     }
-});
+}
+AccountTypeSelection.template = "account.AccountTypeSelection";
 
-fieldRegistry.add("account_type_selection", HierarchySelection);
+registry.category("fields").add("account_type_selection", AccountTypeSelection);
